pet-clinic-web/src/main/java/volha/spring/petclinic/comtrollers/VetsController.java
package volha.spring.petclinic.comtrollers;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import volha.spring.petclinic.services.VetService;

@Controller
public class VetsController {

    private final VetService vetService;

    public VetsController(VetService vetService) {
        this.vetService = vetService;
    }

    @RequestMapping({"/vets/", "/vets/index", "/vets/index.html", "/vets"})
    public String listVets(Model model) {

        model.addAttribute("vets", vetService.findAll());

        return "vets/index";
    }
}
